Add files via upload
The capstone project is a focused approach to attempt a real-life challenge with the learnings from the program. The Goals of the Projects achieved are tagged here. • Computer Vision: Locate the position of inflammation in an image. • Natural Language Processing: Build a classifier that can classify the tickets by analyzing text and Design a ML/DL based chatbot utility which can help the professionals to highlight the safety risk as per the incident description

Skills & Tools Covered
CNN
Text processing
LSTM
Chat Bot